Extract session-end prefix matching into a mixin (WS5 seam)
Move the 18 methods that compare the durable store's message prefix
against replayed messages at session end -- and maintain the
per-session LCM bypass/normal message-prefix fingerprints -- verbatim
out of engine.py into a new PrefixMatchingMixin in prefix_matching.py.
LCMEngine mixes it in before ContextEngine, so `self` still resolves
every reference; call sites, shared state, and method names are
unchanged.

The `_LCM_MESSAGE_PREFIX_FINGERPRINT_LIMIT` constant (used only by
this cluster) moves with it; the now-unused `hashlib` import is
dropped from engine.py. engine.py -344 lines.

Behaviour-neutral: the moved method bodies and the constant are
byte-identical to main (AST-slice compare, 0 diffs). ruff clean;
full suite 1575 passed / 12 xfailed (unchanged from the v0.19.0
baseline).
